Factor out VectorOfVectors stuff to separate modules
