+++ src/lgdo/lh5/_serializers/__init__.py
@@ -7,13 +7,16 @@ from .array import (
     _h5_read_ndarray,
 )
 from .composite import (
-    _h5_read_encoded_array,
     _h5_read_lgdo,
     _h5_read_struct,
     _h5_read_table,
-    _h5_read_vector_of_vectors,
+)
+from .encoded import (
+    _h5_read_array_of_encoded_equalsized_arrays,
+    _h5_read_vector_of_encoded_vectors,
 )
 from .scalar import _h5_read_scalar
+from .vector_of_vectors import _h5_read_vector_of_vectors
 
 __all__ = [
     "_h5_read_lgdo",
@@ -26,4 +29,6 @@ __all__ = [
     "_h5_read_struct",
     "_h5_read_table",
     "_h5_read_scalar",
+    "_h5_read_array_of_encoded_equalsized_arrays",
+    "_h5_read_vector_of_encoded_vectors",
 ]
